Ahora se crea un proyecto de bienvenida para cada usuario registrado. También se han realizado mejoras visuales y arreglado algunos bugs.

+++ Task4allJSF-war/src/java/task4all/bean/AdjuntoBean.java
@@ -119,7 +119,9 @@ public class AdjuntoBean {
         ruta = ruta.concat("Task4allJSF-war" + File.separator + "web" + adjunto.getUrl());
         ruta = crearRutaParaLocalizarArchivo(ruta);
         File f = new File(ruta);
-        f.delete();
+        if(f.exists()) {
+            f.delete();
+        }
 
         return "tarea";
     }
